[CLANG] Fix ARM64 warning cleanup issues
Add missing ARM64 intrinsic declarations and Clang warning gates needed by the build.

Tighten signedness, enum, format, and narrowing conversions reported by Clang across affected user-mode, kernel, WDF, filesystem, and loader code.

Fix the NDIS 6 legacy packet wrapper to use the valid Reserved[0]/Reserved[1] slots instead of indexing past the two-entry Reserved array.

+++ dll/win32/gdiplus/brush.c
@@ -54,7 +54,7 @@ double erf(double x)
     sign = 1;
     if (x < 0)
         sign = -1;
-    x = fabsf(x);
+    x = fabs(x);
 
     /* A & S 7.1.26 */
     t = 1.0/(1.0 + p*x);
